requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==25.1.0
boto3==1.37.3
botocore==1.37.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
fastapi==0.115.8
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-texttospeech==2.25.0
googleapis-common-protos==1.68.0
gotrue==2.11.4
greenlet==3.1.1
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jmespath==1.0.1
multidict==6.1.0
packaging==24.2
postgrest==0.19.3
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
realtime==2.4.0
requests==2.32.3
rsa==4.9
s3transfer==0.11.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
yarl==1.18.3
